requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
SQLAlchemy==2.0.35
alembic==1.13.2
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
python-multipart==0.0.9
jinja2==3.1.4
pandas==2.2.3
numpy==2.1.2
scikit-learn==1.5.2
lightgbm==4.5.0
joblib==1.4.2
python-dotenv==1.0.1
tensorflow-cpu==2.20.0
email-validator==2.2.0
Pillow==10.4.0
reportlab==4.2.5
gunicorn==23.0.0
psycopg2-binary==2.9.9
